cmd/gomobile: resolve symbolic links as a workaround for an Xcode bug
On macOS, a temporary directory starts with /var, which is a symbolic
link to /private/var. And in gomobile, a temporary directory is usually
used as a working directly. Unfortunately, xcodebuild in Xcode 15 seems
to have a bug and might not be able to understand fullpaths with
symbolic links. As a workaround, resolve the path with symbolic links by
filepath.EvalSymlinks.

Fixes golang/go#63141

Change-Id: Ieb684223192fc3e354b79dededaecdb0b25937e6
Reviewed-on: https://go-review.googlesource.com/c/mobile/+/530135
TryBot-Result: Gopher Robot <gobot@golang.org>
Run-TryBot: Hajime Hoshi <hajimehoshi@gmail.com>
Reviewed-by: Bryan Mills <bcmills@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Auto-Submit: Hajime Hoshi <hajimehoshi@gmail.com>