bind: pin go/types to not produce types.Alias types
Bind currently targets the earlier go/types semantics (gotypesalias=0),
and needs to be updated to be compatible with the new (gotypesalias=1).

This is the only way in which the module isn't compatible with Go 1.23
language semantics, so add an explicit godebug line for it with a TODO.
This unblocks moving the module to Go 1.23 (as done in the next CL).

For golang/go#70698.

Change-Id: I59a61defdcf327aed9ccd67b123b1516cd118d9e
Reviewed-on: https://go-review.googlesource.com/c/mobile/+/649381
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Hajime Hoshi <hajimehoshi@gmail.com>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Reviewed-by: Michael Knyszek <mknyszek@google.com>